Навел порядок: убрал @Autowired над конструкторами в Bot и в Logic, пренес сущности в папку entity
