+++ src/main/java/com/example/demo_hh_stat/HhApi.java
@@ -1,5 +1,9 @@
-package com.example.demo_hh_stat.segundo;
+package com.example.demo_hh_stat;
 
+import com.example.demo_hh_stat.entity.Country;
+import com.example.demo_hh_stat.entity.Filter;
+import com.example.demo_hh_stat.entity.ListVacancies;
+import com.example.demo_hh_stat.entity.Vacancy;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.stereotype.Component;
