+++ src/main/java/com/example/demo_hh_stat/entity/ListVacancies.java
@@ -1,4 +1,4 @@
-package com.example.demo_hh_stat.segundo;
+package com.example.demo_hh_stat.entity;
 
 import lombok.Data;
